split all task by (1) (2) (3)
