--- Zhemanov_Alexsand_lesson_2/my_result.yaml
@@ -1,21 +1,0 @@
-sensor:
-- T1
-- T2
-- T3
-average_temperature_celsius: 58
-report:
-  15.09.22: 33℃
-  12.09.22: 63℃
-  06.10.22: 11℃
-  15.10.22: 124℃
-  04.11.22: 70℃
-  12.11.22: 6℃
-  06.12.22: 83℃
-  13.12.22: 50℃
-  22.01.23: 72℃
-  23.01.23: 108℃
-  01.02.23: 72℃
-  07.02.23: 6℃
-  untest: 12℉
-  a: €
-  b: €
