+++ Zhemanov_Alexsand_lesson_2/res/out_csv.csv
@@ -1,0 +1,4 @@
+"Изготовитель системы","Название ОС","Код продукта","Тип системы"
+"LENOVO","Microsoft Windows 7 Профессиональная ","00971-OEM-1982661-00231","x64-based PC"
+"ACER","Microsoft Windows 10 Professional","00971-OEM-1982661-00231","x64-based PC"
+"DELL","Microsoft Windows 8","00971-OEM-1982661-00231","x86-based PC"
