+++ Zhemanov_Alexsand_lesson_2/task_1.py
@@ -12,8 +12,6 @@
  - Создать функцию write_order_to_json(), в которую передается 5 параметров — товар (item), количество (quantity), цена (price), покупатель (buyer), дата (date). Функция должна предусматривать запись данных в виде словаря в файл orders.json. При записи данных указать величину отступа в 4 пробельных символа;
  - Проверить работу программы через вызов функции write_order_to_json() с передачей в нее значений каждого параметра. 
 """
-import os
-from datetime import date
 
 
 def get_data() -> list[list[str]]:
@@ -103,47 +101,10 @@ def write_to_csv(pth: "str"):
         writer.writerows(data)
 
 
-# in poduct price must be decimal
-def write_order_to_json(item: str, quantity: int, price: float, buyer: str,
-                        date: date):
-    """
-    """
-    import json
-
-    from chardet.universaldetector import UniversalDetector
-
-    FILE_PATH = "./res/orders.json"
-    new_order = {
-        "item": item.encode("utf-8").decode("utf-8"),
-        "quantity": quantity,
-        "price": price,
-        "buyer": buyer,
-        "date": date.__str__()
-    }
-
-    # get json encoding
-    detector = UniversalDetector()
-    encoding = "utf-8"
-    with open(FILE_PATH, "rb") as file_h:
-        for line_b in file_h:
-            detector.feed(line_b)
-            if detector.done:
-                encoding = detector.result["encoding"]
-                break
-
-    with open(FILE_PATH, "r+", encoding=encoding) as file_h:
-        cur_data = json.load(file_h)
-        orders = cur_data.get("orders", [])
-        orders.append(new_order)
-        file_h.seek(0)
-        json.dump(cur_data, file_h, indent=4, ensure_ascii=False)
-
-
 if __name__ == "__main__":
     import chardet
 
-    CSV_FILE = "out_csv.csv"
-    JSON_FILE = "./res/orders.json"
+    CSV_FILE = "./res/out_csv.csv"
     # --
     write_to_csv(CSV_FILE)
     print(f"{f'Inside {CSV_FILE}':-^79}")
@@ -153,16 +114,4 @@ if __name__ == "__main__":
         text = bytes_.decode(enc)
         for line in text:
             print(line, end="")
-    print(f"{'':-^79}")
-    # --
-    write_order_to_json("книга", 3, 12.50, "some_buyer", date.today())
-
-    print(f"{f'Inside {JSON_FILE}':-^79}")
-    with open(JSON_FILE, "rb") as f_h:
-        bytes_ = f_h.read()
-        enc = chardet.detect(bytes_).get("encoding") or "utf-8"
-        text = bytes_.decode(enc)
-        for line in text:
-            print(line, end="")
-        print()
     print(f"{'':-^79}")
