+++ Zhemanov_Alexsand_lesson_2/task_2.py
@@ -1,72 +1,80 @@
 """
-3. Задание на закрепление знаний по модулю yaml.
-    Написать скрипт, автоматизирующий сохранение данных в файле YAML-формата.
+    2. Задание на закрепление знаний по модулю json. 
+    Есть файл orders в формате JSON с информацией о заказах.
+    Написать скрипт, автоматизирующий его заполнение данными.
     Для этого:
-        Подготовить данные для записи в виде словаря,
-        в котором первому ключу соответствует список,
-        второму — целое число,
-        третьему — вложенный словарь,
-            где значение каждого ключа — это целое число с юникод-символом,
-        отсутствующим в кодировке ASCII (например, €);
-    
-    Реализовать сохранение данных в файл формата YAML — например, в файл file.yaml.
-    При этом обеспечить стилизацию файла с помощью параметра default_flow_style,
-    а также установить возможность работы с юникодом: allow_unicode = True;
-    
-    Реализовать считывание данных из созданного файла и проверить, совпадают ли они с исходными.
+        - Создать функцию write_order_to_json(),
+        в которую передается 5 параметров — 
+            товар (item),
+            количество (quantity),
+            цена (price),
+            покупатель (buyer),
+            дата (date).
+        Функция должна предусматривать запись данных 
+            в виде словаря в файл orders.json. 
+        При записи данных указать величину отступа в 4 пробельных символа;
+    - Проверить работу программы через вызов функции write_order_to_json() 
+        с передачей в нее значений каждого параметра. 
 """
+from datetime import date
 
-import yaml
+# in poduct price must be decimal
+import json
 
-JSON_ = {
-    "sensor": ["T1", "T2", "T3"],
-    "average_temperature_celsius": 58,
-    "report": {
-        "15.09.22": "33℃",
-        "12.09.22": "63℃",
-        "06.10.22": "11℃",
-        "15.10.22": "124℃",
-        "04.11.22": "70℃",
-        "12.11.22": "6℃",
-        "06.12.22": "83℃",
-        "13.12.22": "50℃",
-        "22.01.23": "72℃",
-        "23.01.23": "108℃",
-        "01.02.23": "72℃",
-        "07.02.23": "6℃",
-        "error": "12℉",
+from chardet.universaldetector import UniversalDetector
+
+
+def write_order_to_json(item: str, quantity: int, price: float, buyer: str,
+                        date: date):
+    """
+    add order to ./res/orders.json
+    """
+
+    FILE_PATH = "./res/orders.json"
+    new_order = {
+        "item": item.encode("utf-8").decode("utf-8"),
+        "quantity": quantity,
+        "price": price,
+        "buyer": buyer,
+        "date": date.__str__()
     }
-}
 
-FILE_PATH = "my_result.yaml"
+    # get json encoding
+    detector = UniversalDetector()
+    encoding = "utf-8"
+    with open(FILE_PATH, "rb") as file_h:
+        for line_b in file_h:
+            detector.feed(line_b)
+            if detector.done:
+                encoding = detector.result["encoding"]
+                break
 
-
-def serization_dict_to_yaml(file_path: str,
-                            data: dict = JSON_,
-                            default_flow_style=False,
-                            allow_unicode=False):
-    """
-    serization python dictionary to file
-
-    param file_path - path to serization
-    param data - python dictionary
-    param default_flow_style - serization list as block
-    param allow_unicode - support unicode charset
-    """
-
-    with open(file_path, "w", encoding="utf-8") as f_h:
-        yaml.dump(data,
-                  f_h,
-                  default_flow_style=default_flow_style,
-                  allow_unicode=allow_unicode,
-                  sort_keys=False)
+    with open(FILE_PATH, "r+", encoding=encoding) as file_h:
+        cur_data = json.load(file_h)
+        orders = cur_data.get("orders", [])
+        orders.append(new_order)
+        file_h.seek(0)
+        json.dump(cur_data, file_h, indent=4, ensure_ascii=False)
 
 
-# serization
-serization_dict_to_yaml(FILE_PATH,
-                        default_flow_style=False,
-                        allow_unicode=True)
-# test
-with open(FILE_PATH, "r", encoding="utf-8") as f_read:
-    res = yaml.safe_load(f_read)
-    assert (JSON_ == res)
+if __name__ == "__main__":
+    import chardet
+
+    JSON_FILE = "./res/orders.json"
+    # --
+    write_order_to_json("книга_1", 3, 120.50, "some_buyer", date.today())
+    write_order_to_json("Все лето в один день", 7, 400.00, "some_buyer",
+                        date.today())
+    write_order_to_json("книга_3", 4, 300.75, "some_buyer", date.today())
+    write_order_to_json("451 ℉", 1, 550.45, "some_buyer", date.today())
+    # --
+
+    print(f"{f'Inside {JSON_FILE}':-^79}")
+    with open(JSON_FILE, "rb") as f_h:
+        bytes_ = f_h.read()
+        enc = chardet.detect(bytes_).get("encoding") or "utf-8"
+        text = bytes_.decode(enc)
+        for line in text:
+            print(line, end="")
+        print()
+    print(f"{'':-^79}")
